requirements.txt
Django==1.7.4
Markdown==2.5.2
PyYAML==3.11
Pygments==2.0.2
contextlib2==0.4.0
coverage==3.7.1
django-filter==0.9.2
django-registration==1.0
djangorestframework==3.0.5
djangorestframework-bulk==0.2
gnureadline==6.3.3
jsonfield==1.0.3
jsonschema==2.4.0
lxml==3.4.2
mock==1.0.1
psycopg2==2.6
python-dateutil==2.4.1
requests==2.5.3
sh==1.11
six==1.9.0
urwid==1.3.0
vcrpy==1.2.0
wrapt==1.10.4
wsgiref==0.1.2
